web/requirements.txt
click==7.1.2
cycler==0.10.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
Werkzeug==1.0.1
wordcloud==1.8.1
